Add Android version of zombicide project. Make changes elsewhere for support. Remove all rendering from ZBoard and all in the UIZBoardComponent for consistency. getTextHeight now returns float instead of int which makes it better for cross-platform.

+++ AndroidGameLibrary/src/cc/lib/android/GL10Graphics.java
@@ -128,11 +128,11 @@ public class GL10Graphics extends AGraphics {
      * 
      * @return
      */
-    public final int getTextHeight() {
+    public final float getTextHeight() {
         return textHeight;
     }
 
-    private int textHeight = 20;
+    private float textHeight = 20;
 
     @Override
     public final void setTextHeight(float height) {
